Move mpi test fixtures into a conftest.py file and add fixture to synchronize states between the processes.
